Backend/requirements.txt
apturl==0.5.2
asgiref==3.3.4
asyncpg==0.23.0
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cmake==3.20.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
databases==0.4.3
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.4
djangorestframework==3.12.4
entrypoints==0.3
fastapi==0.65.2
greenlet==0.4.15
gunicorn==20.1.0
h11==0.12.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
msgpack==0.6.2
mypy==0.761
mypy-extensions==0.4.3
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.9.1
pycairo==1.16.2
pycups==1.9.73
pydantic==1.8.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.24
sqlparse==0.4.1
starlette==0.14.2
systemd-python==234
tornado==6.1
typed-ast==1.4.1
typing-extensions==3.10.0.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.14.0
wadllib==1.3.3
webencodings==0.5.1
xkit==0.0.0
